requirements.txt
discord
python-dotenv
pymongo[srv]
motor